% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hdb_add_data}
\alias{hdb_add_data}
\title{Add Data}
\usage{
hdb_add_data(data, file, resolution = "abort")
}
\arguments{
\item{data}{A data frame of data with columns Station, Date, Hour, Time, Value.
Additional optional columns include Corrected, Status,}

\item{file}{A string of the name of the database file.}

\item{resolution}{A string of the action to take with regard to existing values.
Options are abort, ignore or replace.}
}
\value{
A data frame of the imported parameters.
}
\description{
Add Data
}
