% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{hdb_add_parameters}
\alias{hdb_add_parameters}
\title{Add Parameters}
\usage{
hdb_add_parameters(parameters, file)
}
\arguments{
\item{parameters}{A data frame of parameters with columns Parameter and Units.}

\item{file}{A string of the name of the database file.}
}
\description{
Add Parameters
}
