% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{hdb_add_station}
\alias{hdb_add_station}
\title{Add Station}
\usage{
hdb_add_station(station, parameter, file)
}
\arguments{
\item{station}{A string of the station name.}

\item{parameter}{A string of the parameter.}

\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the imported station.
}
\description{
Add Station
}
