% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{hdb_add_stations}
\alias{hdb_add_stations}
\title{Add Stations}
\usage{
hdb_add_stations(stations, file)
}
\arguments{
\item{stations}{A data frame of stations with columns Parameter and Units.}

\item{file}{A string of the name of the database file.}
}
\value{
The imported stations.
}
\description{
Add Stations
}
