% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{hdb_create}
\alias{hdb_create}
\title{Create Hourly Database}
\usage{
hdb_create(file = "hdb.sqlite")
}
\arguments{
\item{file}{A string of the name of the database file.}
}
\description{
Creates an empty SQLite database to store hourly data.
}
