% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ts_add_data}
\alias{ts_add_data}
\title{Add Data}
\usage{
ts_add_data(data, aggregate = NULL, na_rm = FALSE,
  resolution = "abort", conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{data}{A data frame of data with columns Station, DateTime, Recorded.
Additional optional columns include Corrected, Status, Comments.}

\item{aggregate}{A function to aggregate multiple values within the same station period.}

\item{na_rm}{A flag indicating whether to remove missing values (if possible) when aggregating.}

\item{resolution}{A string of the action to take with regard to existing values.
Options are 'abort', 'ignore' or 'replace'.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the imported parameters.
}
\description{
Times are rounded down prior to import.
}
