% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ts_add_parameter}
\alias{ts_add_parameter}
\title{Add Parameter}
\usage{
ts_add_parameter(parameter, units, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{parameter}{A string of the parameter name.}

\item{units}{A string of the units}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the imported parameters.
}
\description{
Add Parameter
}
