% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{ts_add_station}
\alias{ts_add_station}
\title{Add Station}
\usage{
ts_add_station(station, parameter, period, site,
  file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{station}{A string of the station}

\item{parameter}{A string of the parameter.}

\item{period}{A string of the period. The possible values are 'year', 'month',
'day', 'hour', 'minute' and 'second'.}

\item{site}{A string of the site.}

\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the imported station.
}
\description{
Add Station
}
