% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{ts_add_stations}
\alias{ts_add_stations}
\title{Add Stations}
\usage{
ts_add_stations(stations, file)
}
\arguments{
\item{stations}{A data frame of stations with columns Station, Parameter,
LowerLimit, UpperLimit, Longitude, Latitude, Organization,
StationName.}

\item{file}{A string of the name of the database file.}
}
\value{
The imported stations.
}
\description{
Add Stations
}
