% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ts_connect_db}
\alias{ts_connect_db}
\title{Connect to Database}
\usage{
ts_connect_db(file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{file}{A string of the name of the database file.}
}
\value{
A connection.
}
\description{
Connect to Database
}
