% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{ts_delete_station}
\alias{ts_delete_station}
\title{Delete Station}
\usage{
ts_delete_station(station, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{station}{A string of the station name.}

\item{conn}{An object of class SQLiteConnection.}
}
\description{
Deletes all records associated with a station from the database.
}
