% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ts_disconnect_db}
\alias{ts_disconnect_db}
\title{Disconnect from Database}
\usage{
ts_disconnect_db(conn)
}
\arguments{
\item{conn}{An object of class SQLiteConnection.}
}
\description{
Disconnect from Database
}
