% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ts_get_data}
\alias{ts_get_data}
\title{Get Data}
\usage{
ts_get_data(stations, start_date = as.Date("2017-01-01"),
  end_date = as.Date("2017-12-31"), period = "hour",
  status = "questionable", fill = FALSE, file = getOption("tsdbr.file",
  "ts.db"))
}
\arguments{
\item{stations}{A character vector of the stations.}

\item{start_date}{The start date.}

\item{end_date}{The start date.}

\item{period}{A string of the period to aggregate and average by.
The possible values are 'year', 'month', 'day', 'hour', 'minute' and 'second'.}

\item{status}{A string of the worse type of data to get.
The possible values are 'reasonable', 'questionable' or 'erroneous'.}

\item{fill}{A flag indicating whether to fill in missing values with NAs.}

\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the requested data.
}
\description{
Get Data
}
