% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{ts_get_log}
\alias{ts_get_log}
\title{Get Log Table}
\usage{
ts_get_log(file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the requested data.
}
\description{
Gets Log table as a data frame.
}
