% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintainer.R
\name{ts_get_maintainer}
\alias{ts_get_maintainer}
\title{Get Maintainer}
\usage{
ts_get_maintainer(file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{file}{A string of the name of the database file.}
}
\value{
A string of the maintainer.
}
\description{
Gets maintainer as a string.
}
