% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{ts_get_sites}
\alias{ts_get_sites}
\title{Get Site Table}
\usage{
ts_get_sites(file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the requested data.
}
\description{
Gets site table as a data frame.
}
