% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{ts_get_stations}
\alias{ts_get_stations}
\title{Get Stations Table}
\usage{
ts_get_stations(parameters = NULL, periods = c("year", "month", "day",
  "hour", "minute", "second"), file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{parameters}{A character of the parameters to filter by.}

\item{periods}{A character vector of the periods to filter by.}

\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the requested data.
}
\description{
Gets stations table as a data frame.
}
