% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ts_integer_to_status}
\alias{ts_integer_to_status}
\title{Status to Integer}
\usage{
ts_integer_to_status(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
An ordered factor of status values.
}
\description{
Converts an integer vector to an ordered factor of status values.
}
\examples{
ts_integer_to_status(1:3)
}
