% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ts_plot_data}
\alias{ts_plot_data}
\title{Plot Data}
\usage{
ts_plot_data(data)
}
\arguments{
\item{data}{A data frame of the data to plot with minimum of columns Station,
DateTime and Recorded and/or Corrected.}
}
\description{
Plot Data
}
