% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ts_status_to_integer}
\alias{ts_status_to_integer}
\title{Status to Integer}
\usage{
ts_status_to_integer(x)
}
\arguments{
\item{x}{An ordered factor of status values.}
}
\value{
An integer vector.
}
\description{
Converts an ordered factor of status values to an integer vector.
}
\examples{
ts_status_to_integer(ordered("questionable", 
 c("reasonable", "questionable", "erroneous")))
}
