% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ts_utc_offset_to_tz}
\alias{ts_utc_offset_to_tz}
\title{UTC Offset to Time Zone}
\usage{
ts_utc_offset_to_tz(utc_offset)
}
\arguments{
\item{utc_offset}{An integer of the UTC offset.}
}
\value{
The time zone as a string.
}
\description{
UTC Offset to Time Zone
}
\examples{
ts_utc_offset_to_tz(-8L)
}
