% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ts_add_parameter}
\alias{ts_add_parameter}
\title{Add Parameter}
\usage{
ts_add_parameter(parameter, units, file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{parameter}{A string of the parameter name.}

\item{units}{A string of the units}

\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the imported parameters.
}
\description{
Add Parameter
}
