% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ts_add_parameters}
\alias{ts_add_parameters}
\title{Add Parameters}
\usage{
ts_add_parameters(parameters, file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{parameters}{A data frame of parameters with columns Parameter and Units.}

\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the imported parameters.
}
\description{
Add Parameters
}
