% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{ts_add_site}
\alias{ts_add_site}
\title{Add Sites}
\usage{
ts_add_site(site, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{site}{A string of the site name.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the imported parameters.
}
\description{
Add Sites
}
