% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{ts_add_site}
\alias{ts_add_site}
\title{Add Sites}
\usage{
ts_add_site(site, file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{site}{A string of the site name.}

\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the imported parameters.
}
\description{
Add Sites
}
