% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{ts_add_sites}
\alias{ts_add_sites}
\title{Add Site}
\usage{
ts_add_sites(sites, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{sites}{A data frame of parameters with columns Site.
The optional columns are
Longitude, Latitude, Organization, SiteName and Comments.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the imported parameters.
}
\description{
Add Site
}
