% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{ts_add_stations}
\alias{ts_add_stations}
\title{Add Stations}
\usage{
ts_add_stations(stations, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{stations}{A data frame of stations with columns Station, Parameter,
Period and Site. The optional columns are
Depth, LowerLimit, UpperLimit, StationName, StationID and Comments.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
The imported station data.
}
\description{
Add Stations
}
