% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{ts_create}
\alias{ts_create}
\title{Create Time-Series Database}
\usage{
ts_create(file = "ts.db", utc_offset = 0L)
}
\arguments{
\item{file}{A string of the name of the database file.}

\item{utc_offset}{A integer of the utc offset which must lie between -12 and 14.}
}
\description{
Creates an empty SQLite database to store time series data.
The utc_offset indicates how many hours must be added or subtracted to the 
saved date times to convert them to UTC.
For example 8 hours must be added to times in PST to convert them to UTC.
}
