% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ts_get_data}
\alias{ts_get_data}
\title{Get Data}
\usage{
ts_get_data(stations = NULL, start_date = NULL, end_date = NULL,
  period = "hour", na_rm = FALSE, status = "questionable",
  fill = FALSE, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{stations}{A character vector of the stations.}

\item{start_date}{The start date.}

\item{end_date}{The end date.}

\item{period}{A string of the period to aggregate and average by.
The possible values are 'year', 'month', 'day', 'hour', 'minute' and 'second'.}

\item{na_rm}{A flag indicating whether to remove missing values (if possible) when aggregating.}

\item{status}{A string of the worse type of data to get.
The possible values are 'reasonable', 'questionable' or 'erroneous'.}

\item{fill}{A flag indicating whether to fill in gaps (with missing values).}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the requested data.
}
\description{
Get Data
}
