% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclaimer.R
\name{ts_get_disclaimer}
\alias{ts_get_disclaimer}
\title{Get Disclaimer}
\usage{
ts_get_disclaimer(file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{file}{A string of the name of the database file.}
}
\value{
A string of the disclaimer.
}
\description{
Gets disclaimer as a string.
}
