% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{ts_get_log}
\alias{ts_get_log}
\title{Get Log Table}
\usage{
ts_get_log(conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the requested data.
}
\description{
Gets Log table as a data frame.
}
