% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ts_get_parameters}
\alias{ts_get_parameters}
\title{Get Parameter Table}
\usage{
ts_get_parameters(file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the requested data.
}
\description{
Gets parameter table as a data frame.
}
