% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{ts_get_sites}
\alias{ts_get_sites}
\title{Get Site Table}
\usage{
ts_get_sites(conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the requested data.
}
\description{
Gets site table as a data frame.
}
