% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ts_get_table}
\alias{ts_get_table}
\title{Get Table}
\usage{
ts_get_table(table, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{table}{A string of the table to get.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the table
}
\description{
Get Table
}
