% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclaimer.R
\name{ts_set_disclaimer}
\alias{ts_set_disclaimer}
\title{Set Disclaimer}
\usage{
ts_set_disclaimer(
  disclaimer = "THE DATA ARE COPYRIGHTED",
  conn = getOption("tsdbr.conn", NULL)
)
}
\arguments{
\item{disclaimer}{A string of the disclaimer.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A string of the old disclaimer.
}
\description{
Sets disclaimer.
}
