% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{ts_translate_stations}
\alias{ts_translate_stations}
\title{Translate Stations}
\usage{
ts_translate_stations(data, from = "StationID", to = "Station",
  na_rm = TRUE, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{data}{A data frame with the column Station of the station coding.}

\item{from}{A string indicating the coding to translate from. 
Possible values are 'Station', 'StationName' and 'StationID'.}

\item{to}{A string indicating the coding to translate to.}

\item{na_rm}{A flag indicating whether to remove missing stations.
Possible values are 'Station', 'StationName' and 'StationID'.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
The translated data
}
\description{
Translate stations from or to the codings in Station, StationName and StationID.
Its useful if the data are provided with different station codings to the main ones used in the database.
}
